+++ application_sdk/io/__init__.py
@@ -29,6 +29,7 @@ from temporalio import activity
 from application_sdk.activities.common.models import ActivityStatistics
 from application_sdk.activities.common.utils import get_object_store_prefix
 from application_sdk.common.types import DataframeType
+from application_sdk.constants import ENABLE_ATLAN_UPLOAD, UPSTREAM_OBJECT_STORE_NAME
 from application_sdk.io._utils import (
     estimate_dataframe_record_size,
     is_empty_dataframe,
@@ -342,6 +343,13 @@ class Writer(ABC):
 
     async def _upload_file(self, file_name: str):
         """Upload a file to the object store."""
+        if ENABLE_ATLAN_UPLOAD:
+            await ObjectStore.upload_file(
+                source=file_name,
+                store_name=UPSTREAM_OBJECT_STORE_NAME,
+                retain_local_copy=True,
+                destination=get_object_store_prefix(file_name),
+            )
         await ObjectStore.upload_file(
             source=file_name,
             destination=get_object_store_prefix(file_name),